added rudimentary backend

+++ models/positions.js
@@ -2,7 +2,7 @@ const positions = (connection, Sequelize) => {
   return connection.define('positions', {
     id: { type: Sequelize.INTEGER, autoIncrement: true, primaryKey: true },
     position: { type: Sequelize.STRING, allowNull: false },
-    salary: { type: Sequelize.INTEGER, allowNull: false },
+    salary: { type: Sequelize.DECIMAL, allowNull: false },
     benefits: { type: Sequelize.INTEGER, allowNull: false },
   }, {
     defaultScope: { attributes: { exclude: ['deletedAt'] } }
